BUG-100624 Update Cloudbreak Mock
