+++ mock/cloudbreak/controllers/Info.js
@@ -11,7 +11,3 @@ module.exports.getCloudbreakInfo = function getCloudbreakInfo (req, res, next) {
 module.exports.getCloudbreakHealth = function getCloudbreakHealth (req, res, next) {
     Info.getCloudbreakHealth(req.swagger.params, res, next);
 };
-
-module.exports.getPeriscopeHealth = function getPeriscopeHealth (req, res, next) {
-    Info.getPeriscopeHealth(req.swagger.params, res, next);
-};
