+++ mock/periscope/api/swagger.yaml
@@ -10,8 +10,9 @@ info:
     \ level QoS to the cluster."
   version: "1.9.0"
   title: "Auto-scaling API"
-basePath: "/as/api"
+basePath: "/as"
 tags:
+- name: "health"
 - name: "v1alerts"
 - name: "v2clusters"
 - name: "v1history"
@@ -22,7 +23,28 @@ schemes:
 - "http"
 - "https"
 paths:
-  /v1/clusters:
+  /health:
+    get:
+      tags:
+      - "health"
+      summary: "retrieve Periscope server status for user"
+      description: "Periscope server status."
+      operationId: "getPeriscopeHealth"
+      schemes:
+      - "http"
+      - "https"
+      consumes:
+      - "application/json"
+      produces:
+      - "application/json"
+      parameters: []
+      responses:
+        200:
+          description: "successful operation"
+          schema:
+            $ref: "#/definitions/Health"
+      x-swagger-router-controller: "Health"
+  /api/v1/clusters:
     get:
       tags:
       - "v1clusters"
@@ -70,7 +92,7 @@ paths:
           schema:
             $ref: "#/definitions/AutoscaleClusterResponse"
       x-swagger-router-controller: "V1clusters"
-  /v1/clusters/{clusterId}:
+  /api/v1/clusters/{clusterId}:
     get:
       tags:
       - "v1clusters"
@@ -149,7 +171,7 @@ paths:
         default:
           description: "successful operation"
       x-swagger-router-controller: "V1clusters"
-  /v1/clusters/{clusterId}/alerts/metric:
+  /api/v1/clusters/{clusterId}/alerts/metric:
     get:
       tags:
       - "v1alerts"
@@ -219,7 +241,7 @@ paths:
           schema:
             $ref: "#/definitions/MetricAlertResponse"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/metric/definitions:
+  /api/v1/clusters/{clusterId}/alerts/metric/definitions:
     get:
       tags:
       - "v1alerts"
@@ -256,7 +278,7 @@ paths:
                 type: "object"
                 properties: {}
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/metric/{alertId}:
+  /api/v1/clusters/{clusterId}/alerts/metric/{alertId}:
     put:
       tags:
       - "v1alerts"
@@ -332,7 +354,7 @@ paths:
         default:
           description: "successful operation"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/prometheus:
+  /api/v1/clusters/{clusterId}/alerts/prometheus:
     get:
       tags:
       - "v1alerts"
@@ -390,7 +412,7 @@ paths:
           schema:
             $ref: "#/definitions/PrometheusAlertResponse"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/prometheus/definitions:
+  /api/v1/clusters/{clusterId}/alerts/prometheus/definitions:
     get:
       tags:
       - "v1alerts"
@@ -418,7 +440,7 @@ paths:
             items:
               $ref: "#/definitions/AlertRuleDefinitionEntry"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/prometheus/{alertId}:
+  /api/v1/clusters/{clusterId}/alerts/prometheus/{alertId}:
     put:
       tags:
       - "v1alerts"
@@ -482,7 +504,7 @@ paths:
         default:
           description: "successful operation"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/time:
+  /api/v1/clusters/{clusterId}/alerts/time:
     get:
       tags:
       - "v1alerts"
@@ -544,7 +566,7 @@ paths:
           schema:
             $ref: "#/definitions/TimeAlertResponse"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/time/validate:
+  /api/v1/clusters/{clusterId}/alerts/time/validate:
     post:
       tags:
       - "v1alerts"
@@ -577,7 +599,7 @@ paths:
           schema:
             type: "boolean"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/alerts/time/{alertId}:
+  /api/v1/clusters/{clusterId}/alerts/time/{alertId}:
     put:
       tags:
       - "v1alerts"
@@ -645,7 +667,7 @@ paths:
         default:
           description: "successful operation"
       x-swagger-router-controller: "V1alerts"
-  /v1/clusters/{clusterId}/autoscale:
+  /api/v1/clusters/{clusterId}/autoscale:
     post:
       tags:
       - "v1clusters"
@@ -676,7 +698,7 @@ paths:
           schema:
             $ref: "#/definitions/AutoscaleClusterResponse"
       x-swagger-router-controller: "V1clusters"
-  /v1/clusters/{clusterId}/configurations/scaling:
+  /api/v1/clusters/{clusterId}/configurations/scaling:
     get:
       tags:
       - "v1configurations"
@@ -746,7 +768,7 @@ paths:
           schema:
             $ref: "#/definitions/ScalingConfiguration"
       x-swagger-router-controller: "V1configurations"
-  /v1/clusters/{clusterId}/history:
+  /api/v1/clusters/{clusterId}/history:
     get:
       tags:
       - "v1history"
@@ -774,7 +796,7 @@ paths:
             items:
               $ref: "#/definitions/HistoryResponse"
       x-swagger-router-controller: "V1history"
-  /v1/clusters/{clusterId}/history/{historyId}:
+  /api/v1/clusters/{clusterId}/history/{historyId}:
     get:
       tags:
       - "v1history"
@@ -805,7 +827,7 @@ paths:
           schema:
             $ref: "#/definitions/HistoryResponse"
       x-swagger-router-controller: "V1history"
-  /v1/clusters/{clusterId}/policies:
+  /api/v1/clusters/{clusterId}/policies:
     get:
       tags:
       - "v1policies"
@@ -875,7 +897,7 @@ paths:
           schema:
             $ref: "#/definitions/ScalingPolicyResponse"
       x-swagger-router-controller: "V1policies"
-  /v1/clusters/{clusterId}/policies/{policyId}:
+  /api/v1/clusters/{clusterId}/policies/{policyId}:
     put:
       tags:
       - "v1policies"
@@ -951,7 +973,7 @@ paths:
         default:
           description: "successful operation"
       x-swagger-router-controller: "V1policies"
-  /v1/clusters/{clusterId}/state:
+  /api/v1/clusters/{clusterId}/state:
     post:
       tags:
       - "v1clusters"
@@ -982,7 +1004,7 @@ paths:
           schema:
             $ref: "#/definitions/AutoscaleClusterResponse"
       x-swagger-router-controller: "V1clusters"
-  /v2/clusters/{cbClusterId}:
+  /api/v2/clusters/{cbClusterId}:
     get:
       tags:
       - "v2clusters"
@@ -1061,7 +1083,7 @@ paths:
         default:
           description: "successful operation"
       x-swagger-router-controller: "V2clusters"
-  /v2/clusters/{cbClusterId}/disable:
+  /api/v2/clusters/{cbClusterId}/disable:
     post:
       tags:
       - "v2clusters"
@@ -1087,7 +1109,7 @@ paths:
           schema:
             $ref: "#/definitions/AutoscaleClusterResponse"
       x-swagger-router-controller: "V2clusters"
-  /v2/clusters/{cbClusterId}/enable:
+  /api/v2/clusters/{cbClusterId}/enable:
     post:
       tags:
       - "v2clusters"
@@ -1113,7 +1135,7 @@ paths:
           schema:
             $ref: "#/definitions/AutoscaleClusterResponse"
       x-swagger-router-controller: "V2clusters"
-  /v2/clusters/{cbClusterId}/running:
+  /api/v2/clusters/{cbClusterId}/running:
     post:
       tags:
       - "v2clusters"
@@ -1139,7 +1161,7 @@ paths:
           schema:
             $ref: "#/definitions/AutoscaleClusterResponse"
       x-swagger-router-controller: "V2clusters"
-  /v2/clusters/{cbClusterId}/suspended:
+  /api/v2/clusters/{cbClusterId}/suspended:
     post:
       tags:
       - "v2clusters"
@@ -1607,3 +1629,5 @@ definitions:
       cronExpression:
         type: "string"
         description: "Cron expression of the time alert"
+  Health:
+    type: "object"
